+++ internal/request/containers_test.go
@@ -4,7 +4,7 @@
 package request_test
 
 import (
-	request "github.com/gowizzard/mobyspulse/internal/requests"
+	"github.com/gowizzard/mobyspulse/internal/request"
 	"reflect"
 	"testing"
 )
