+++ internal/request/main_test.go
@@ -1,0 +1,11 @@
+package request_test
+
+import (
+	"github.com/gowizzard/mobyspulse/internal/request"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	m.Run()
+	request.Client.CloseIdleConnections()
+}
